requirements.txt
python-can==4.2.2
flask==2.3.2
flask_cors==4.0.0
pygame==2.5.2
requests
google-cloud-pubsub==2.23.0